% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase-write.R
\name{add_color}
\alias{add_color}
\title{Add a named color to an ASE swatch object}
\usage{
add_color(ase_obj, name, model, color_vals, type)
}
\arguments{
\item{ase_obj}{An ASE object created with \code{\link[=create_ase]{create_ase()}}}

\item{name}{name of the color}

\item{model}{one of "\code{RGB}", "\code{Gray}", "\code{CMYK}", or "\code{LAB}"}

\item{color_vals}{vector of color values associated with the chosen \code{model}}

\item{type}{one of "\code{global}", "\code{spot}", or "\code{process}"}
}
\value{
\code{ase_obj} (invisibly)
}
\description{
Add a named color to an ASE swatch object
}
\examples{
create_ase() |>
add_color(
  name = "RGB Red",
  model = "RGB",
  color_vals = as.vector(col2rgb("#FF0000")/255),
  type = "global"
) |>
  add_color(
    name = "RGB Yellow",
    model = "RGB",
    color_vals = as.vector(col2rgb("#FFFF00")/255),
    type = "global"
  ) |>
  ase_encode() |>
  writeBin(tempfile(fileext = ".ase"))
}
