% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase-write.R
\name{ase_encode}
\alias{ase_encode}
\title{Encode a built ASE object for output}
\usage{
ase_encode(data)
}
\arguments{
\item{data}{An ASE swatch object created by \code{\link[=create_ase]{create_ase()}}}
}
\value{
raw vector
}
\description{
Encode a built ASE object for output
}
\examples{
create_ase() |>
add_color(
  name = "RGB Red",
  model = "RGB",
  color_vals = as.vector(col2rgb("#FF0000")/255),
  type = "global"
) |>
  add_color(
    name = "RGB Yellow",
    model = "RGB",
    color_vals = as.vector(col2rgb("#FFFF00")/255),
    type = "global"
  ) |>
  ase_encode() |>
  writeBin(tempfile(fileext = ".ase"))
}
