% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase-write.R
\name{create_ase}
\alias{create_ase}
\title{Create an ASE object}
\usage{
create_ase()
}
\value{
a \code{list} (invisibly) classed as \code{ase}
}
\description{
Create an ASE object
}
\examples{
create_ase() |>
add_color(
  name = "RGB Red",
  model = "RGB",
  color_vals = as.vector(col2rgb("#FF0000")/255),
  type = "global"
) |>
  add_color(
    name = "RGB Yellow",
    model = "RGB",
    color_vals = as.vector(col2rgb("#FFFF00")/255),
    type = "global"
  ) |>
  ase_encode() |>
  writeBin(tempfile(fileext = ".ase"))
}
