% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_swatch.R
\name{plot_swatch}
\alias{plot_swatch}
\title{Plot a color swatch from hex}
\usage{
plot_swatch(palette, nrow = 1)
}
\arguments{
\item{palette}{A character vector with colors in hex}

\item{nrow}{Number of rows to display the palette in}
}
\value{
A ggplot object
}
\description{
Plot a color swatch from hex
}
\examples{
x <- c("#FF0000", "#00FFFF", "#0000FF")
plot_swatch(x)
}
