% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wheel.R
\name{plot_wheel}
\alias{plot_wheel}
\title{Plot a color palette on color wheel}
\usage{
plot_wheel(palette)
}
\arguments{
\item{palette}{A character vector with colors in hex}
}
\value{
A ggplot object
}
\description{
Plot a color palette on color wheel
}
\examples{
x <- c("#FF0000", "#00FFFF", "#0000FF")
plot_wheel(x)
}
