% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swatches.R
\name{show_palette}
\alias{show_palette}
\title{Display a color palette}
\usage{
show_palette(palette)
}
\arguments{
\item{palette}{vector of character hex RGB values}
}
\description{
Given a character vector (hex RGB values), display palette in graphics window.
}
\examples{
# built-in palette
keep_the_change <- read_ase(system.file("palettes",
                            "keep_the_change.ase", package="swatches"))
print(keep_the_change)
# show_palette(keep_the_change)
}

