% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches-package.R
\docType{package}
\name{swatches}
\alias{swatches}
\alias{swatches-package}
\title{Read, Inspect, Manipulate, and Save Color Swatch Files}
\description{
There are numerous places to create and download color
palettes. These are usually shared in Adobe swatch file formats of
some kind. There is also often the need to use standard palettes
developed within an organization to ensure that aesthetics are carried over
into all projects and output. Now there is a way to read these swatch
files in R and avoid transcribing or converting color values by hand or
or with other programs. This package provides functions to read and
inspect Adobe Color (ACO), Adobe Swatch Exchange (ASE), GIMP Palette (GPL),
OpenOffice palette (SOC) files and KDE Palette ("colors") files. Detailed
descriptions of Adobe Color and Swatch Exchange file formats as well as other
swatch file formats can be found at \url{http://www.selapa.net/swatches/colors/fileformats.php}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/swatches}
  \item Report bugs at \url{https://github.com/hrbrmstr/swatches/issues}
}

}
\author{
\strong{Maintainer}: Bob Rudis \email{bob@rud.is} (\href{https://orcid.org/0000-0001-5670-2640}{ORCID})

}
\keyword{internal}
