% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-palette.R
\name{trim_palette}
\alias{trim_palette}
\title{Try to intelligently reduce a large palette down to a reasonable smaller
set of colors}
\usage{
trim_palette(pal, n = 5)
}
\arguments{
\item{pal}{input palette to reduct}

\item{n}{number of desired colors}
}
\value{
vector of \code{n} colors from \code{pal}
}
\description{
Given a palette and a desired number of colors to use, this function will
compute CIEDE2000 and attempt to reduce the input set to a disctinct
smaller set of colors based on color distances.
}
\note{
internal CIEDE2000 color distance implementation by Gaurav Sharma & Maynard P Baalthazar
}
